Rename(#56): 패키지 분리
