+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/domain/Schedule.java
@@ -1,6 +1,6 @@
-package com.wypl.jpacalendardomain.calendar.domain;
+package com.wypl.jpacalendardomain.schedule.domain;
 
-import com.wypl.jpacalendardomain.calendar.data.RepetitionCycle;
+import com.wypl.jpacalendardomain.schedule.data.RepetitionCycle;
 import com.wypl.jpacommon.JpaBaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
