+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/domain/ScheduleInfo.java
@@ -1,5 +1,6 @@
-package com.wypl.jpacalendardomain.calendar.domain;
+package com.wypl.jpacalendardomain.schedule.domain;
 
+import com.wypl.jpacalendardomain.calendar.domain.Calendar;
 import com.wypl.jpacommon.JpaBaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
