+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/repository/ScheduleInfoRepository.java
@@ -1,6 +1,6 @@
-package com.wypl.jpacalendardomain.calendar.repository;
+package com.wypl.jpacalendardomain.schedule.repository;
 
-import com.wypl.jpacalendardomain.calendar.domain.ScheduleInfo;
+import com.wypl.jpacalendardomain.schedule.domain.ScheduleInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ScheduleInfoRepository extends JpaRepository<ScheduleInfo, Long> {
